Let there be modules.
Moved the larger types into their own modules to
try to keep things organised and make repetition
a bit more obvious.

Removed as much template haskell as I could, unsure
if I should have waited until later in the design process
for this to occur. Oh well, shame Haskell is so hard to
refactor, eh? eh?

Cleaned up the Example.lhs a bit.

Shaved some of the sharper edges off the use of some
of the iso/prisms.

Added a more robust AnimDuration type.

Changed the props functions to use a more lensy style.
Happy with it for now, and it seems to be highlighting
something that I could clean up into a more useful type.