Add ViewBox Attribute to SVG_El.
** Breaking Change **
Allow setting of the 'viewBox' on the SVG element, this was missing before and
is a very very, very, VERY, useful component.

Expose the looser SVG element drawing functions. This package does not have
coverage over all possible SVG elements so best not to prevent people from
creating their own extensions.

Updated the 'default.nix' to prevent Haddocks from being generated with a GHCJS
build. This was causing build failures in packages that depended on this one.

Added a function that allows you to change the indexed type of a Pos.